test(canon): split the reachability tests into a sibling file
dependency_scan.rs crossed the new-file source budget once the
declaration-file and alias-pruning coverage landed; the module keeps the
pass and its helpers, the tests move next door.
